config: Add a config for use with cloud init
To set up all the apt-available tools and sources for building
uZAP Open Source projects (ZapDragon LLVM and ZapPi Kernel).

This isn't very portable with all the download steps & full
configuration, so just set up the smaller, non-git parts for the
build environment.

Usage for a fresh env should probably be documented and/or
improved on, but currently don't have the bandwidth, so the source
was just transferred via mounting and copying locally so these steps
were superceded for me.

Signed-off-by: notcarbide <luis@0k.cx>